Varios ajustes
-Que la tabla de pags este en la basdats
-Que se genere correctamente el jar por medio de maven junto con todas las dependencias correctamente: Falta que empaque el archivo de hibernate.config.xml adentro del

empacado
-En el modulo de ventas al hacer una nueva venta que no permita agregar productos hasta que no este agregado el header, esto por la moneda que debe de tomar cuando se

selecciona un producto
-En el modulo de ventas al hacer una nueva venta cuando se seleccina un producto que tome la moneda del header
-En el modulo de ventas al hacer una nueva venta cuando se seleccina un producto que seleccione por default la unidad del producto
-En el modulo de ventas al hacer una nueva venta cuando se seleccina un producto que seleccione por default el almacén de ventas
-Cuando se crea el sistema por primera vez que cree el almacén de ventas
-En el repositorio de almacenes que regrese el almacen de ventas por medio de un método
-En el modulo de ventas al hacer una nueva venta cuando se seleccino un producto me quita el almacén de ventas que estaba seleccionado previamente
-En el modulo de ventas cuando genero una venta no esta respetando el almacén que estoy seleccionando para la existencia del almacén del producto
-Cuando se intenta borrar un almacén de los catálogos de almacénes y este ya tiene movimientos que no permita borrarlo
-En el modulo de ventas que comience con el cliente mostrador al cargarse para eficientizar la venta
-Integrar la pantalla para ver texto o la descripción de un producto en grande
-En el modulo de ventas al hacer una nueva venta cuando se seleccina un producto que la pantalla de ver descripcion en grande funcione

+++ src/main/java/com/era/easyretail/controllers/views/AlmasViewController.java
@@ -9,6 +9,7 @@ import com.era.datamodels.enums.SearchCommonTypeEnum;
 import com.era.easyretail.validators.ValidatorFactory;
 import com.era.easyretail.validators.WarehousesValidator;
 import com.era.logger.LoggerUtility;
+import com.era.models.Existalma;
 import com.era.models.User;
 import com.era.models.Warehouse;
 import com.era.repositories.RepositoryFactory;
@@ -344,14 +345,21 @@ public class AlmasViewController extends AlmasJFrame {
                 });
                 return ;
             }
+    
+            //Get the selected model
+            final Warehouse Warehouse = (Warehouse)jTab.getRowSelected();
+                    
+            //If the warehouse is assign to any product stop
+            final Existalma Existalma = RepositoryFactory.getInstance().getExistalmasRepository().getByWarehouse(Warehouse.getCode());
+            if(Existalma!=null){
+                DialogsFactory.getSingleton().showErrorOKCallbackDialog(baseJFrame, "errors_warehouse_cannot_delete", null);
+                return;
+            }
             
             //Question if continue
             DialogsFactory.getSingleton().showQuestionContinueDialog(baseJFrame, (JFrame jFrame) -> {
                 
                 try{
-                
-                    //Get the selected model
-                    final Warehouse Warehouse = (Warehouse)jTab.getRowSelected();
                     
                     //Delete the record
                     RepositoryFactory.getInstance().getWarehousesRepository().delete(Warehouse);
